requirements.txt
Flask~=1.1.2
gunicorn==20.0.4
pymessenger~=0.0.7.0
flask-redis 
python-dotenv
pylint
pytest
fasttext~=0.9.2
tensorflow~=2.2.0
numpy~=1.18.5
keras~=2.3.1
spacy~=2.3.0
pyspellchecker
googletrans~=3.0.0
pandas~=1.0.4
fakeredis~=1.4.1
azure-cognitiveservices-search-websearch
redis~=3.5.3
msrest~=0.6.16
setuptools~=44.0.0
dropbox
